Update ng-scrollable.js
Change lines 261-273 where scrollSpy events were broadcast to now be emit. This was causing the events not to be able to be heard by the controller - there was a problem with the isolated scope of the directive. See this stackoverflow question for the exact details: http://stackoverflow.com/questions/42326919/how-to-use-spymargin-with-ng-scrollable/42328321